Sorted Builder functions into script interpretation, analysis and script generation.
